+++ src/guillo-gmi/components/panel/items.tsx
@@ -20,12 +20,7 @@ import { Input } from '../input/input'
 import { SelectVocabulary } from '../input/select_vocabulary'
 import { useIntl } from 'react-intl'
 import { genericMessages } from '../../locales/generic_messages'
-
-import {
-  ItemColumn,
-  RegistrySchemaFilter,
-  SearchItem,
-} from '../../types/guillotina'
+import { RegistrySchemaFilter, SearchItem } from '../../types/guillotina'
 
 interface InitialState {
   page: number
@@ -120,11 +115,12 @@ export function PanelItems() {
     if (Ctx.state.loading) return
     const getData = async () => {
       setState({ loading: true, total: Ctx.context.length })
-      const { get } = Ctx.registry
-      const fnName = get('searchEngineQueryParamsFunction', SearchEngine)
+      const { getSearchEngineQueryParamsFunction } = Ctx.registry
+      const fnName = getSearchEngineQueryParamsFunction(SearchEngine)
       if (sortParsed === undefined) {
         const defaultSortValue = Ctx.registry.getDefaultSortValue(
-          Ctx.context['@type']
+          Ctx.context['@type'],
+          undefined
         )
         sortParsed = parser(
           `_sort_${defaultSortValue.direction}=${defaultSortValue.key}`
@@ -339,7 +335,7 @@ export function PanelItems() {
               <th>
                 <AllItemsCheckbox />
               </th>
-              {columns.map((column: ItemColumn) => (
+              {columns.map((column) => (
                 <th
                   key={`table-col-${column.label}`}
                   onClick={() => column.isSortable && onSort(column.key)}