+++ src/guillo-gmi/hooks/useRegistry.tsx
@@ -13,7 +13,7 @@ import { RemoveItems } from '../actions/remove_items'
 import { RemoveItem } from '../actions/remove_item'
 import { AddItem } from '../actions/add_item'
 import { ChangePassword } from '../actions/change_pass'
-import { BaseForm } from '../forms/base'
+import { BaseForm, BaseFormProps } from '../forms/base'
 import { UserForm } from '../forms/users'
 import { IAttachment } from '../components/behaviors/iattachment'
 import { IDublinCore } from '../components/behaviors/idublincore'
@@ -35,13 +35,24 @@ import {
   ItemColumn,
   RegistrySchema,
 } from '../types/guillotina'
+import { buildQs, parser } from '../lib/search'
 
+export interface RegistrySortValue {
+  direction: 'asc' | 'des'
+  key: string
+}
+export interface RegistryProperties {
+  Buttons: React.ReactElement
+  Panels: React.ReactElement
+  default: string[]
+  ignoreField: string[]
+}
 export interface IRegistry {
   paths: {
     [key: string]: React.FC
   }
   views: {
-    [key: string]: React.FC | React.ComponentType<any>
+    [key: string]: React.ComponentType<any>
   }
   actions: {
     [key: string]: (props: any) => JSX.Element
@@ -59,7 +70,7 @@ export interface IRegistry {
     [key: string]: RegistrySchema
   }
   properties: {
-    [key: string]: React.FC
+    [key: string]: RegistryProperties
   }
   components: {
     [key: string]: (props: any) => React.ReactNode | null | undefined
@@ -70,13 +81,14 @@ export interface IRegistry {
   fieldsToFilter: {
     [key: string]: string[]
   }
+  parseSearchQueryParamFunction: {
+    [key: string]: (query: string, type: string) => string
+  }
   defaultSortValue: {
-    [key: string]: {
-      direction: 'asc' | 'des'
-      field: string
-    }
+    [key: string]: RegistrySortValue
   }
 }
+
 const registry: IRegistry = {
   paths: {},
   views: {
@@ -135,20 +147,40 @@ const registry: IRegistry = {
   fieldsToFilter: {
     UserManager: ['id', 'email', 'user_name'],
   },
+  parseSearchQueryParamFunction: {},
   defaultSortValue: {},
 }
 
-const get = (key: keyof IRegistry, param: string, fallback = undefined) => {
-  if (registry[key]) return registry[key][param] || fallback
-  return fallback
+export interface IManageRegistry {
+  registry: IRegistry
+  getPathComponent: (
+    context: GuillotinaCommonObject | undefined,
+    path: string,
+    fallback?: React.FC
+  ) => React.ComponentType<any>
+  getComponent: (name: string) => React.ComponentType<any>
+  getView: (name: string) => React.ComponentType<any>
+  getForm: (type: string, fallback?: React.FC) => React.FC<BaseFormProps>
+  getAction: (type: string, fallback?: React.FC) => React.FC
+  getBehavior: (type: string, fallback?: React.FC) => React.FC<any>
+  getProperties: (type: string) => RegistryProperties
+  getItemsColumn: (type: string) => ItemColumn[] | undefined
+  getSchemas: (type: string) => RegistrySchema
+  getFieldsToFilter: (type: string, fallback?: string[]) => string[]
+  getParsedSearchQueryParam: (query: string, type: string) => string
+  getDefaultSortValue: (
+    type: string,
+    fallback?: RegistrySortValue
+  ) => RegistrySortValue
+  getSearchEngineQueryParamsFunction: (type: string) => string
 }
 
-const getComponent = (
+const getPathComponent = (
   context: GuillotinaCommonObject | undefined,
   path: string,
-  fallback = undefined
+  fallback: React.FC | undefined = undefined
 ) => {
-  if (!context) return
+  if (!context) return () => null
   // console.log("Component for path", path)
   // lookup by path
   if (registry.paths[path]) {
@@ -172,11 +204,15 @@ const getItemsColumn = (type: string) => {
   return undefined
 }
 
+const getComponent = (name: string) => {
+  return registry.components[name]
+}
+
 const getView = (name: string) => {
   return registry.views[name]
 }
 
-const getForm = (type: string, fallback: React.FC) => {
+const getForm = (type: string, fallback?: React.FC<BaseFormProps>) => {
   return registry.forms[type] || fallback || BaseForm
 }
 
@@ -184,7 +220,7 @@ const getAction = (type: string, fallback?: React.FC) => {
   return registry.actions[type] || fallback
 }
 
-const getBehavior = (type: string, fallback: React.FC) => {
+const getBehavior = (type: string, fallback?: React.FC) => {
   return registry.behaviors[type] || fallback
 }
 
@@ -202,7 +238,7 @@ const getFieldsToFilter = (type: string, fallback: string[] = ['title']) => {
 
 const getDefaultSortValue = (
   type: string,
-  fallback = {
+  fallback: RegistrySortValue | undefined = {
     key: 'id',
     direction: 'des',
   }
@@ -210,11 +246,24 @@ const getDefaultSortValue = (
   return registry.defaultSortValue[type] || fallback
 }
 
+const getSearchEngineQueryParamsFunction = (type: string) => {
+  return registry.searchEngineQueryParamsFunction[type]
+}
+
+const getParsedSearchQueryParam = (query: string, type: string) => {
+  const parsedFunction = registry.parseSearchQueryParamFunction[type]
+  if (!parsedFunction) {
+    const fieldsToFilter = getFieldsToFilter(type)
+    return buildQs(parser(query, fieldsToFilter))
+  }
+  return parsedFunction(query, type)
+}
+
 export const defaultComponent = (context: GuillotinaCommonObject) => {
   return context.is_folderish ? FolderCtx : ItemCtx
 }
 
-export function useRegistry(data: IRegistry) {
+export function useRegistry(data: Partial<IRegistry>): IManageRegistry {
   // if data is provided we need to merge it into actual registry
   const ref = React.useRef<unknown>()
   if (data && !ref.current) {
@@ -230,17 +279,19 @@ export function useRegistry(data: IRegistry) {
 
   return {
     registry,
-    get,
     getForm,
     getComponent,
+    getPathComponent,
     getAction,
     getBehavior,
     getProperties,
     getItemsColumn,
     getFieldsToFilter,
+    getParsedSearchQueryParam,
     getDefaultSortValue,
     getSchemas,
     getView,
+    getSearchEngineQueryParamsFunction,
   }
 }